Rephrase non-goal 4 (Action #336) (#341)
* Rephrase non-goal 4 (Action #336)

Addresses #336

We agreed that we need implementations to demonstrate success. The original non-goal 4 said we would not create implementations. Technically this is true--the MFWG is not creating them, implementers are. However, hewing to the gold standard of "rough consensus and running code", we recognize the need for implementations to exist. To address this, I made a new deliverable 6 defining the need for (but not the direct WG production of) interoperable implementations.

In addition, I did some light editing throughout. Notably:

- defined "dynamic message string", as "localizable message string" is a more expansive definition actually beyond our scope (static messages are out of scope)

* Address @eemeli's comments

* Fixed the typo