Update examples to indent 'when' lines by one space
I'd like to suggest that, when formatting messages over multiple lines, the _recommended_ formatting for `when` clauses be (a) each one on its own line, and (b) indented by one space. This hopefully can help discern the "logic" parts of complex messages (`let`, `match`) from the "content" parts (`when`).

    let $count = {$count :number fractionDigits=0}
    match {$count :plural}
     when 1 {You have one notification.}
     when * {You have {$count} notifications.}

I'm opening a PR to demonstrate what it would look like.