Propose reserving additional sigils for future use (#374)
* Update message.abnf

This proposal proposes to reserve sigils for future use with reserved sigils allowed in the same place that function names are currently, e.g. standalone (`{%foo}`), with options (`{%foo key=val}`), or subsidiary to variable/literals (`{|quote me| %foo}` or `{$var %foo}`).

A case could be made for reserving a comment syntax that allows a placeholder with prose in it (and I would suggest using the `%` for it). @eemeli's proposal is effectively this. I didn't propose this because of the impact on parsers in the future. Let's discuss.

* Update spec/message.abnf

I think this works. I'm going to commit it to the PR and then do some additional cleanup.

Co-authored-by: Eemeli Aro <eemeli@mozilla.com>

* Cleanup

- align some items
- move `function` with `variable`
- move `reserve-escape` with the other escapes

* Fix typo

* Add `reserved` to the spec

Note that this change also removed the undefined term "operand" in favor of just using literal and variable (which are defined terms). It also imposes 2119 keywords.

* allow passing values to reserved

* Add warning about whitespace significance in `reserved`

* Fix expression links, add more about reserved

* Address comments

* Make spaces insignificant at the end of reserved sequences

* Remove the warning about trailing whitespace on reserved

* Fix reserved to ensure no ending whitespace

---------

Co-authored-by: Eemeli Aro <eemeli@mozilla.com>