Change the syntax of the ::open and :/close function calls
This is a follow-up to #364, which made it possible to use unquoted literals in the argument position in placeholders. However, due to the current syntax of +open and -close function calls, arguments that are number literals must still be quoted, e.g. `{|-1| :number}`.

This PR proposes to change the syntax of markup-like function calls:

    BEFORE: {+button title=|Click me!|}Submit{-button}
    AFTER:  {::button title=|Click me!|}Submit{:/button}

The benefit of using a two-char-long prefix is that we effectively establish the colon `:` as the general-purpose function introducer.