(registry.dtd) Rename pattern to validationRule (#435)
We already use `pattern` to describe a sequence of text and placeholders in the message body. Let's not overload it with another, validation-related meaning. I had picked `pattern` in `registry.dtd` because of "regex patterns", but I think we can do better than that.

This PR also changes some `NMTOKEN` attributes to `IDREF`, as noticed by @eemeli in https://github.com/unicode-org/message-format-wg/issues/407#issuecomment-1621606634.

This PR doesn't address the main question of #407, however. It's only concerned with renaming `pattern` to prevent confusion.