Add normative language about expression nesting
Fixes #353 

We previously agreed not to document not allowing function/expression nesting. This is a minor fix to add that. The ABNF already doesn't permit nesting, so it's really just documentation cleanup. 

In addition, I added examples of variables as options in two places as an oversight in our set of examples.