Replace 'shape' with better terminology
Addresses #590

Replaces the amorphous term "shape" in Formatting with more appropriate text. Generally I used a reformulation using some variation of the word "form".

In one case, I deleted the paragraph. The paragraph in question was non-normative and basically said the same thing as a normative paragraph just below it--that how formatting works is implementation-defined.