Replace 'shape' with better terminology (#611)
* Replace 'shape' with better terminology

Addresses #590

Replaces the amorphous term "shape" in Formatting with more appropriate text. Generally I used a reformulation using some variation of the word "form".

In one case, I deleted the paragraph. The paragraph in question was non-normative and basically said the same thing as a normative paragraph just below it--that how formatting works is implementation-defined.

* Fix note formatting and make normative ("important")

The note about eager vs. lazy was intended to be normative. Putting it in a note makes it informative. I changed to "important" and fixed it to format correctly in markdown.