Add bidi isolate support
This commit allows bidi isolates as optional characters in whitespace (both required and optional whitespace).

It also explicitly permits decoration of `placeholder` and `quoted-pattern` with LRI/PDI on the correct (non-text interfering) side of the marker. This meant moving the `{`/`}` marker to `placeholder`.

I did not add decoration to `literal` because `owsp` and `wsp` take care of permitting the controls wherever that production is used. As near as I can tell, all tokens can be decorated with isolates or LRM/RLM