Clean up `quoted-pattern` definition
`complex-body` was already always prefixed by optional space, so in that case there was effectively two subsequent optional spaces: that is unnecessary. It is much cleaner to remove the optional leading space from the `quoted-pattern` definition, and instead move it to explicitly before `quoted-pattern` in  `variant`.

This change is purely editorial and has no normative impact.